refactor tournament editing
- moves all edit operations to multiple service methods instead of having a gigancting controller edit method
